refactor: keep only provider-agnostic transforms in internal/jsonschema
Per review feedback: anything named after a provider does not belong in a
provider-agnostic package. Move OpenAI's keyword allowlist,
stripUnsupportedKeywords and the base64-hint helper into providers/openai as
unexported code; they call the now-exported jsonschema.Walk for traversal.

internal/jsonschema keeps CollapseDynamicNodes (used by both OpenAI and
Anthropic), exposes Walk as a reusable tree-walker, and gains DeepCopy, which
replaces the deepCopyMap that was copy-pasted in the openai and anthropic
mappers.
